Refactor:삭제 모듈화, edit 모듈화
