+++ init.js
@@ -1,4 +1,4 @@
-import {loadToDos} from "./todoApp.js"
+import { loadToDos } from "./todoApp.js"
 function init() {
     console.log("load toDos");
     loadToDos();
